Javdoc + Visibility + More Cleanup (#229)
