+++ automatedtests/src/androidTestR2_16_1/java/com/mux/stats/sdk/muxstats/automatedtests/ui/SimplePlayerTestActivity.java
@@ -37,10 +37,11 @@ import com.google.android.exoplayer2.upstream.DataSpec;
 import com.google.android.exoplayer2.upstream.DefaultDataSourceFactory;
 import com.google.android.exoplayer2.util.Util;
 import com.mux.stats.sdk.muxstats.AdsImaSDKListener;
+import com.mux.stats.sdk.muxstats.SimplePlayerBaseActivity;
 import com.mux.stats.sdk.muxstats.automatedtests.R;
 
 
-public class SimplePlayerTestActivity extends SimplePlayerBaseActivity  implements Player.Listener{
+public class SimplePlayerTestActivity extends SimplePlayerBaseActivity implements Player.Listener{
 
   MediaSourceFactory mediaSourceFactory;
