+++ automatedtests/src/androidTestR2_9_6/java/com/mux/stats/sdk/muxstats/automatedtests/ui/SimplePlayerTestActivity.java
@@ -28,6 +28,7 @@ import com.google.android.exoplayer2.upstream.DefaultDataSourceFactory;
 import com.google.android.exoplayer2.util.Util;
 import com.mux.stats.sdk.muxstats.automatedtests.R;
 import java.lang.reflect.Constructor;
+import com.mux.stats.sdk.muxstats.SimplePlayerBaseActivity;
 
 public class SimplePlayerTestActivity extends SimplePlayerBaseActivity
     implements Player.EventListener {
